--- a/.gitignore
+++ b/.gitignore
@@ -11,6 +11,7 @@ __pycache__/
 
 # Testing and coverage
 .coverage
+coverage.json
 htmlcov/
 .ruff_cache/