--- a/src/finances/core/currency.py
+++ b/src/finances/core/currency.py
@@ -61,10 +61,11 @@ def cents_to_dollars_str(cents: int) -> str:
         cents: Amount in cents
 
     Returns:
-        Formatted dollar string
+        Formatted dollar string with comma separators
 
     Example:
         cents_to_dollars_str(4599) -> "45.99"
+        cents_to_dollars_str(10000000) -> "100,000.00"
     """
     # Handle negative amounts properly
     is_negative = cents < 0
@@ -73,10 +74,13 @@ def cents_to_dollars_str(cents: int) -> str:
     dollars = int(abs_cents // 100)
     remainder = int(abs_cents % 100)
 
+    # Add comma separators to dollars part
+    dollars_str = f"{dollars:,}"
+
     if is_negative:
-        return f"-{dollars}.{remainder:02d}"
+        return f"-{dollars_str}.{remainder:02d}"
     else:
-        return f"{dollars}.{remainder:02d}"
+        return f"{dollars_str}.{remainder:02d}"
 
 
 def safe_currency_to_cents(currency_str: str | int | float) -> int: