… .pre-commit-config.yaml
… CLAUDE.md
… KNOWN_ISSUES.md
… TEST_COVERAGE_GAPS.md
… coverage.json
… dev/todos.md
… tests/README.md
… tests/e2e/test_apple_cli.py
… tests/e2e/test_flow_system.py
… tests/e2e/test_ynab_cli.py
… tests/fixtures/e2e_helpers.py
… tests/fixtures/synthetic_data.py
… tests/integration/test_amazon_cli_parameters.py
… tests/integration/test_apple_cli_parameters.py
… tests/integration/test_cashflow_cli_parameters.py
… tests/integration/test_flow_integration.py
… tests/integration/test_ynab_cli_parameters.py
… tests/test_data/README.md
… tests/test_data/amazon/sample_orders.csv
… tests/test_data/apple/sample_receipt.html
… tests/test_data/ynab/accounts.json
… tests/test_data/ynab/categories.json
… tests/test_data/ynab/transactions.json
… tests/unit/test_cli/test_flow_cli.py
… tests/unit/test_core/test_currency.py
… tests/unit/test_flow/test_flow_core.py
… tests/unit/test_flow/test_flow_engine.py
… tests/unit/test_ynab/test_retirement.py